fix: update the library to use solc-wrapper
