+++ solc-wrapper/src/ast/parse.rs
@@ -1,7 +1,7 @@
-use crate::ast::ast::Ast;
+use crate::ast::ast::SourceUnit;
 
 use super::error::AstError;
 
-pub fn parse_ast(json: &str) -> Result<Ast, AstError> {
+pub fn parse_ast(json: &str) -> Result<SourceUnit, AstError> {
     Ok(serde_json::from_str(json)?)
 }
