+++ solidhunter-lib/src/linter.rs
@@ -4,13 +4,10 @@ use crate::rules::factory::RuleFactory;
 use crate::rules::rule_impl::{create_rules_file, parse_rules};
 
 use glob::glob;
-use solc_wrapper::Solc;
-
-struct Ast;
-
+use solc_wrapper::{Solc, SourceUnit};
 
 pub struct SolidFile {
-    pub data: Ast,
+    pub data: SourceUnit,
     pub path: String,
     pub content: String,
 }
@@ -59,7 +56,7 @@ impl SolidLinter {
         false
     }
 
-    fn update_file_ast(&mut self, path: &str, ast: Ast) {
+    fn update_file_ast(&mut self, path: &str, ast: SourceUnit) {
         for file in &mut self.files {
             if file.path == path {
                 file.data = ast.clone();
@@ -67,7 +64,7 @@ impl SolidLinter {
         }
     }
 
-    fn add_file(&mut self, path: &str, ast: Ast, content: &str) {
+    fn add_file(&mut self, path: &str, ast: SourceUnit, content: &str) {
         let file = SolidFile {
             data: ast,
             path: String::from(path),
@@ -77,8 +74,7 @@ impl SolidLinter {
     }
 
     pub fn parse_file(&mut self, filepath: String) -> LintResult{
-        let ast = Solc::default().execute_on_file(filepath.as_str()).unwrap();
-        let res = Solc::parse(ast).unwrap();
+        let res = Solc::default().extract_ast_file(filepath.clone()).unwrap();
 
         if self.file_exists(filepath.as_str()) {
             self.update_file_ast(filepath.as_str(), res);
@@ -92,31 +88,32 @@ impl SolidLinter {
             res.append(&mut diags);
         }
         LintResult {
-            errors: res.into_iter().filter(|x| x.severity.unwrap() == Severity::ERROR).collect(),
-            warnings: res.into_iter().filter(|x| x.severity.unwrap() == Severity::WARNING).collect(),
-            infos: res.into_iter().filter(|x| x.severity.unwrap() == Severity::INFO).collect(),
-            hints: res.into_iter().filter(|x| x.severity.unwrap() == Severity::HINT).collect(),
+            errors: res.clone().into_iter().filter(|x| x.severity.unwrap() == Severity::ERROR).collect(),
+            warnings: res.clone().into_iter().filter(|x| x.severity.unwrap() == Severity::WARNING).collect(),
+            infos: res.clone().into_iter().filter(|x| x.severity.unwrap() == Severity::INFO).collect(),
+            hints: res.clone().into_iter().filter(|x| x.severity.unwrap() == Severity::HINT).collect(),
         }
     }
 
-    pub fn parse_content(&mut self, filepath: String, content : String) -> LintResult{
-        let res = Solc::default().extract_ast_content(content).unwrap();
+    pub fn parse_content(&mut self, filepath: String, content : &String) -> LintResult{
+        let res = Solc::default().extract_ast_content(content.to_string()).unwrap();
 
         if self.file_exists(filepath.as_str()) {
             self.update_file_ast(filepath.as_str(), res);
         } else {
             self.add_file(filepath.as_str(), res, content.as_str());
         }
+        let mut res : Vec<LintDiag> = Vec::new();
 
         for rule in &self.rules {
             let mut diags = rule.diagnose(&self.files[0], &self.files);
             res.append(&mut diags);
         }
         LintResult {
-            errors: res.into_iter().filter(|x| x.severity.unwrap() == Severity::ERROR).collect(),
-            warnings: res.into_iter().filter(|x| x.severity.unwrap() == Severity::WARNING).collect(),
-            infos: res.into_iter().filter(|x| x.severity.unwrap() == Severity::INFO).collect(),
-            hints: res.into_iter().filter(|x| x.severity.unwrap() == Severity::HINT).collect(),
+            errors: res.clone().into_iter().filter(|x| x.severity.unwrap() == Severity::ERROR).collect(),
+            warnings: res.clone().into_iter().filter(|x| x.severity.unwrap() == Severity::WARNING).collect(),
+            infos: res.clone().into_iter().filter(|x| x.severity.unwrap() == Severity::INFO).collect(),
+            hints: res.clone().into_iter().filter(|x| x.severity.unwrap() == Severity::HINT).collect(),
         }
     }
 
@@ -131,9 +128,15 @@ impl SolidLinter {
         result
     }
     pub fn delete_file(&mut self, path: String) {
-        for file in &mut self.files {
-            if file.path == path {
-                self.files.remove(self.files.iter().position(|x| x.path == path).unwrap());
+        loop {
+            let idx = self.files.iter().position(|x| x.path == path);
+            match idx {
+                Some(idx) => {
+                    self.files.remove(idx);
+                }
+                None => {
+                    break;
+                }
             }
         }
     }
